refactor: or-1747 modify BeheerVerenigingDetailProjection to EventProjection and split out methods
