+++ src/AssociationRegistry.Admin.Api/Projections/Detail/BeheerVerenigingDetailProjector.cs
@@ -1,0 +1,153 @@
+namespace AssociationRegistry.Admin.Api.Projections.Detail;
+
+using System.Linq;
+using Constants;
+using Events;
+using Framework;
+using Infrastructure.Extensions;
+using Marten.Events;
+
+public class BeheerVerenigingDetailProjector
+{
+    public static void Apply(IEvent<NaamWerdGewijzigd> naamWerdGewijzigd, BeheerVerenigingDetailDocument document)
+    {
+        document.Naam = naamWerdGewijzigd.Data.Naam;
+        document.DatumLaatsteAanpassing = naamWerdGewijzigd.GetHeaderInstant(MetadataHeaderNames.Tijdstip).ToBelgianDate();
+        document.Metadata = new Metadata(naamWerdGewijzigd.Sequence, naamWerdGewijzigd.Version);
+    }
+
+    public static void Apply(IEvent<KorteNaamWerdGewijzigd> korteNaamWerdGewijzigd, BeheerVerenigingDetailDocument document)
+    {
+        document.KorteNaam = korteNaamWerdGewijzigd.Data.KorteNaam;
+        document.DatumLaatsteAanpassing = korteNaamWerdGewijzigd.GetHeaderInstant(MetadataHeaderNames.Tijdstip).ToBelgianDate();
+        document.Metadata = new Metadata(korteNaamWerdGewijzigd.Sequence, korteNaamWerdGewijzigd.Version);
+    }
+
+    public static void Apply(IEvent<KorteBeschrijvingWerdGewijzigd> korteBeschrijvingWerdGewijzigd, BeheerVerenigingDetailDocument document)
+    {
+        document.KorteBeschrijving = korteBeschrijvingWerdGewijzigd.Data.KorteBeschrijving;
+        document.DatumLaatsteAanpassing = korteBeschrijvingWerdGewijzigd.GetHeaderInstant(MetadataHeaderNames.Tijdstip).ToBelgianDate();
+        document.Metadata = new Metadata(korteBeschrijvingWerdGewijzigd.Sequence, korteBeschrijvingWerdGewijzigd.Version);
+    }
+
+    public static void Apply(IEvent<StartdatumWerdGewijzigd> startdatumWerdGewijzigd, BeheerVerenigingDetailDocument document)
+    {
+        document.Startdatum = startdatumWerdGewijzigd.Data.Startdatum?.ToString(WellknownFormats.DateOnly);
+        document.DatumLaatsteAanpassing = startdatumWerdGewijzigd.GetHeaderInstant(MetadataHeaderNames.Tijdstip).ToBelgianDate();
+        document.Metadata = new Metadata(startdatumWerdGewijzigd.Sequence, startdatumWerdGewijzigd.Version);
+    }
+
+    public static void Apply(IEvent<ContactgegevenWerdToegevoegd> contactgegevenWerdToegevoegd, BeheerVerenigingDetailDocument document)
+    {
+        document.Contactgegevens = document.Contactgegevens.Append(
+            new BeheerVerenigingDetailDocument.Contactgegeven
+            {
+                ContactgegevenId = contactgegevenWerdToegevoegd.Data.ContactgegevenId,
+                Type = contactgegevenWerdToegevoegd.Data.Type,
+                Waarde = contactgegevenWerdToegevoegd.Data.Waarde,
+                Beschrijving = contactgegevenWerdToegevoegd.Data.Beschrijving,
+                IsPrimair = contactgegevenWerdToegevoegd.Data.IsPrimair,
+            }).ToArray();
+        document.DatumLaatsteAanpassing = contactgegevenWerdToegevoegd.GetHeaderInstant(MetadataHeaderNames.Tijdstip).ToBelgianDate();
+        document.Metadata = new Metadata(contactgegevenWerdToegevoegd.Sequence, contactgegevenWerdToegevoegd.Version);
+    }
+
+    public static void Apply(IEvent<ContactgegevenWerdGewijzigd> contactgegevenWerdGewijzigd, BeheerVerenigingDetailDocument document)
+    {
+        document.Contactgegevens = document.Contactgegevens
+            .Where(c => c.ContactgegevenId != contactgegevenWerdGewijzigd.Data.ContactgegevenId)
+            .Append(
+                new BeheerVerenigingDetailDocument.Contactgegeven
+                {
+                    ContactgegevenId = contactgegevenWerdGewijzigd.Data.ContactgegevenId,
+                    Type = contactgegevenWerdGewijzigd.Data.Type,
+                    Waarde = contactgegevenWerdGewijzigd.Data.Waarde,
+                    Beschrijving = contactgegevenWerdGewijzigd.Data.Beschrijving,
+                    IsPrimair = contactgegevenWerdGewijzigd.Data.IsPrimair,
+                })
+            .ToArray();
+
+        document.DatumLaatsteAanpassing = contactgegevenWerdGewijzigd.GetHeaderInstant(MetadataHeaderNames.Tijdstip).ToBelgianDate();
+        document.Metadata = new Metadata(contactgegevenWerdGewijzigd.Sequence, contactgegevenWerdGewijzigd.Version);
+    }
+
+    public static void Apply(IEvent<ContactgegevenWerdVerwijderd> contactgegevenWerdVerwijderd, BeheerVerenigingDetailDocument document)
+    {
+        document.Contactgegevens = document.Contactgegevens
+            .Where(
+                c => c.ContactgegevenId != contactgegevenWerdVerwijderd.Data.ContactgegevenId)
+            .ToArray();
+
+        document.DatumLaatsteAanpassing = contactgegevenWerdVerwijderd.GetHeaderInstant(MetadataHeaderNames.Tijdstip).ToBelgianDate();
+        document.Metadata = new Metadata(contactgegevenWerdVerwijderd.Sequence, contactgegevenWerdVerwijderd.Version);
+    }
+
+    public static void Apply(IEvent<HoofdactiviteitenVerenigingsloketWerdenGewijzigd> hoofactiviteitenVerenigingloketWerdenGewijzigd, BeheerVerenigingDetailDocument document)
+    {
+        document.HoofdactiviteitenVerenigingsloket = hoofactiviteitenVerenigingloketWerdenGewijzigd.Data.HoofdactiviteitenVerenigingsloket.Select(
+            h => new BeheerVerenigingDetailDocument.HoofdactiviteitVerenigingsloket
+            {
+                Code = h.Code,
+                Beschrijving = h.Beschrijving,
+            }).ToArray();
+        document.DatumLaatsteAanpassing = hoofactiviteitenVerenigingloketWerdenGewijzigd.GetHeaderInstant(MetadataHeaderNames.Tijdstip).ToBelgianDate();
+        document.Metadata = new Metadata(hoofactiviteitenVerenigingloketWerdenGewijzigd.Sequence, hoofactiviteitenVerenigingloketWerdenGewijzigd.Version);
+    }
+
+    public static void Apply(IEvent<VertegenwoordigerWerdToegevoegd> vertegenwoordigerWerdToegevoegd, BeheerVerenigingDetailDocument document)
+    {
+        document.Vertegenwoordigers = document.Vertegenwoordigers.Append(
+            new BeheerVerenigingDetailDocument.Vertegenwoordiger
+            {
+                VertegenwoordigerId = vertegenwoordigerWerdToegevoegd.Data.VertegenwoordigerId,
+                Achternaam = vertegenwoordigerWerdToegevoegd.Data.Achternaam,
+                Voornaam = vertegenwoordigerWerdToegevoegd.Data.Voornaam,
+                Roepnaam = vertegenwoordigerWerdToegevoegd.Data.Roepnaam,
+                Rol = vertegenwoordigerWerdToegevoegd.Data.Rol,
+                IsPrimair = vertegenwoordigerWerdToegevoegd.Data.IsPrimair,
+                Email = vertegenwoordigerWerdToegevoegd.Data.Email,
+                Telefoon = vertegenwoordigerWerdToegevoegd.Data.Telefoon,
+                Mobiel = vertegenwoordigerWerdToegevoegd.Data.Mobiel,
+                SocialMedia = vertegenwoordigerWerdToegevoegd.Data.SocialMedia,
+            }).ToArray();
+
+        document.DatumLaatsteAanpassing = vertegenwoordigerWerdToegevoegd.GetHeaderInstant(MetadataHeaderNames.Tijdstip).ToBelgianDate();
+        document.Metadata = new Metadata(vertegenwoordigerWerdToegevoegd.Sequence, vertegenwoordigerWerdToegevoegd.Version);
+    }
+
+    public static void Apply(IEvent<VertegenwoordigerWerdGewijzigd> vertegenwoordigerWerdGewijzigd, BeheerVerenigingDetailDocument document)
+    {
+        var vertegenwoordigerToUpdate = document.Vertegenwoordigers.Single(v => v.VertegenwoordigerId == vertegenwoordigerWerdGewijzigd.Data.VertegenwoordigerId);
+        document.Vertegenwoordigers = document.Vertegenwoordigers
+            .Where(c => c.VertegenwoordigerId != vertegenwoordigerWerdGewijzigd.Data.VertegenwoordigerId)
+            .Append(
+                new BeheerVerenigingDetailDocument.Vertegenwoordiger
+                {
+                    VertegenwoordigerId = vertegenwoordigerWerdGewijzigd.Data.VertegenwoordigerId,
+                    Achternaam = vertegenwoordigerToUpdate.Achternaam,
+                    Voornaam = vertegenwoordigerToUpdate.Voornaam,
+                    Roepnaam = vertegenwoordigerWerdGewijzigd.Data.Roepnaam,
+                    Rol = vertegenwoordigerWerdGewijzigd.Data.Rol,
+                    IsPrimair = vertegenwoordigerWerdGewijzigd.Data.IsPrimair,
+                    Email = vertegenwoordigerWerdGewijzigd.Data.Email,
+                    Telefoon = vertegenwoordigerWerdGewijzigd.Data.Telefoon,
+                    Mobiel = vertegenwoordigerWerdGewijzigd.Data.Mobiel,
+                    SocialMedia = vertegenwoordigerWerdGewijzigd.Data.SocialMedia,
+                })
+            .ToArray();
+
+        document.DatumLaatsteAanpassing = vertegenwoordigerWerdGewijzigd.GetHeaderInstant(MetadataHeaderNames.Tijdstip).ToBelgianDate();
+        document.Metadata = new Metadata(vertegenwoordigerWerdGewijzigd.Sequence, vertegenwoordigerWerdGewijzigd.Version);
+    }
+
+    public static void Apply(IEvent<VertegenwoordigerWerdVerwijderd> vertegenwoordigerWerdVerwijderd, BeheerVerenigingDetailDocument document)
+    {
+        document.Vertegenwoordigers = document.Vertegenwoordigers
+            .Where(
+                c => c.VertegenwoordigerId != vertegenwoordigerWerdVerwijderd.Data.VertegenwoordigerId)
+            .ToArray();
+
+        document.DatumLaatsteAanpassing = vertegenwoordigerWerdVerwijderd.GetHeaderInstant(MetadataHeaderNames.Tijdstip).ToBelgianDate();
+        document.Metadata = new Metadata(vertegenwoordigerWerdVerwijderd.Sequence, vertegenwoordigerWerdVerwijderd.Version);
+    }
+}
